refactor: use AutoTestLintOffsets in remaining AnalysisRuleTest suites
