+++ test/src/lints/named_parameters_ordering/named_parameters_ordering_rule_test.dart
@@ -4,7 +4,7 @@ import 'package:solid_lints/src/common/parameter_parser/analysis_options_loader.
 import 'package:solid_lints/src/lints/named_parameters_ordering/named_parameters_ordering_rule.dart';
 import 'package:test_reflective_loader/test_reflective_loader.dart';
 
-import '../../../lints/auto_test_lint_offsets.dart';
+import '../../utils/auto_test_lint_offsets.dart';
 
 void main() {
   defineRefSuite();
@@ -45,9 +45,6 @@ plugins:
       analysisOptionsContent(rules: [rule.name]),
     );
   }
-
-  @override
-  String get analysisRule => NamedParametersOrderingRule.lintName;
 
   Future<void> test_does_not_report_correct_constructor_ordering() async {
     await assertNoDiagnostics(r'''
@@ -259,8 +256,7 @@ void example({
 ''');
   }
 
-  Future<void>
-  test_reports_incorrect_ordering_with_trailing_comments() async {
+  Future<void> test_reports_incorrect_ordering_with_trailing_comments() async {
     await assertAutoDiagnostics('''
 void example({
   int? age, // the age
@@ -272,8 +268,7 @@ void example({
 ''');
   }
 
-  Future<void>
-  test_reports_incorrect_ordering_with_mixed_comments() async {
+  Future<void> test_reports_incorrect_ordering_with_mixed_comments() async {
     await assertAutoDiagnostics('''
 void example({
   /// The age of the user.
@@ -300,8 +295,7 @@ void example({
 ''');
   }
 
-  Future<void>
-  test_does_not_report_with_partial_custom_config() async {
+  Future<void> test_does_not_report_with_partial_custom_config() async {
     newAnalysisOptionsYamlFile(testPackageRootPath, '''
 ${analysisOptionsContent(rules: [rule.name])}
 plugins:
