+++ test/src/utils/auto_lint_data.dart
@@ -1,0 +1,17 @@
+import 'package:analyzer_testing/src/analysis_rule/pub_package_resolution.dart';
+
+class AutoLintData {
+  final String code;
+  final Pattern? correctionContains;
+  final List<Pattern> messageContainsAll;
+  final String? name;
+  final List<ExpectedContextMessage>? contextMessages;
+
+  const AutoLintData({
+    required this.code,
+    this.correctionContains,
+    this.messageContainsAll = const [],
+    this.name,
+    this.contextMessages,
+  });
+}
